pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "order-block"
version = "2.0.0"
description = "AI-powered image quality assessment, similarity clustering, and best-pick selection with modern web UI"
requires-python = ">=3.9"
dependencies = [
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
clustering = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "scikit-learn>=1.3.0",
    "imagehash>=4.3.0",
]
ui = [
    "streamlit>=1.28.0",
]
gui = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "pydantic-settings>=2.0.0",
    "websockets>=12.0",
]
all = [
    "order-block[clustering,gui]",
]
heic = [
    "pillow-heif>=0.13.0",
]

[project.scripts]
order-block = "order_block.cli:main"
order-block-gui = "server.main:start_server"

[tool.setuptools.packages.find]
include = ["order_block*", "server*"]
